chore: move common SQL query code to separate file to avoid duplication
